Removed template for a real script: Dual Bloom
Even inlcudes previews for the readme and a passable amount of in-code
comments. Currently don't know how to set up RunMode.NONINTERACTIVE and
am using a hacky preview method.
I would also like to optionally color the bloom based on its source
hue/saturation, but I'm not sure of how to do that in gegl yet, as it
doesn't appear to have channel-specific composites like HSV H or LCH C.