Convert dual-bloom.py to use bsz_gimp_lib
holy shit it's so much nicer. Especially the param classes. When I first
tested ParamCombo in the UI I thought "wait will this work properly with
reset and preview and blahdeblah and the answer was yes everything
worked first try since I managed to set up the abstract methods
properly.