Rewrote preview functionality
Preview now operates *on the layer* with a backup buffer stored for
'clearing'. This is faster and allows 'persistant' previews via not
flushing during processing.

This is a BREAKING CHANGE, as now bsz_gimp_lib has a simple
"gegl_preview" arg that lets the app know the main function is a
'gegl op', aka works by modifying the current drawable's buffer.

Not quite ready for a new release. Gonna revise Pixel Math and work out
edge cases for the new preview system, as I already expect a few.