Pixel Math 2 Multithreading, refactor for future
Expect a speedup of about 100% with a modern computer. Now stores pixels
in a separate buffer before committing all at once, paving the way for
area operations in the future. This change also firmly cements it as a
specialty tool, not to replace simple plugins like Filmic Chroma as it
is currently a bit of a memory hog using the 2nd buffer. I feel there's
a decent chunk more optimization available that I'm not knowledgeable
enough to implement yet but time will tell if that's true or not.