Reduced scope of PM2 to unbuffered ops only
So no saving all the changes to a vec then mem copying everything later.
The only use case would be area operations, which upon doing a lot of
thinking, is not useful in the slightest. The way it'd be implemented in
the proper 'modular' way would be to write new op code for relative
pixels, example R[-1, 5] would be the red channel 1 left and 5 down. To
do anything useful with this would require basically generating like 100
ops in a python script, which would tank performance to the max as this
library can't make many assumptions/optimizations about the input ops
without writing a full algebra parser.

Benefits:
 - like a 50% speedup just avoiding all the extra copying
 - reduce memory usage by 50%. Running a very large image used like full
     4 gigs just for the internal buffer. It's basically another shadow
     buffer on top of Gegl's.
 - significant code simplicity improvement. might make easier for SIMD
 later.
   - won't need to quadrouple parse_ops() for area ops

Detriments:
 - if you wanted to write the shittiest blur algorithm ever now you
 can't.
 - ???

 I feel these kinds of complex operations would be way better as a
 gegl/gmic/something plugin, not a "I wanna type some maths in a box
 real quick" python plugin. PM2 would be better with other more scalable
 additions instead, like more ops, a 'rand' constant, simple "> < =="
 comparisons, etc.