PM2 vars now init outside loop & box slice
So last night @ 4 am while I was in sleep paralysis I suddenly realized
that I placed the var vec allocation in the loop, so it was creating the
vec form scratch every single pixel. This made no difference in code
which doesn't use vars, but with var heavy code like the last two
benches it made it upwards of *twice as slow*, so I changed that!

also now a boxed slice instead of a vec for smaller pointers. Very very
mild speedup but a speedup nonetheless.